polish up pr
